[property-holder-preparation] Extract BeanMetaDataBuilder to upper level class and create corresponding analog for property holders
