+++ engine/src/main/java/org/hibernate/validator/internal/metadata/aggregated/BeanMetaDataImpl.java
@@ -27,31 +27,17 @@ import javax.validation.metadata.BeanDescriptor;
 import javax.validation.metadata.ConstructorDescriptor;
 import javax.validation.metadata.PropertyDescriptor;
 
-import org.hibernate.validator.internal.engine.MethodValidationConfiguration;
 import org.hibernate.validator.internal.engine.groups.Sequence;
 import org.hibernate.validator.internal.engine.groups.ValidationOrder;
 import org.hibernate.validator.internal.engine.groups.ValidationOrderGenerator;
-import org.hibernate.validator.internal.engine.valueextraction.ValueExtractorManager;
-import org.hibernate.validator.internal.metadata.core.ConstraintHelper;
 import org.hibernate.validator.internal.metadata.core.MetaConstraint;
 import org.hibernate.validator.internal.metadata.descriptor.BeanDescriptorImpl;
 import org.hibernate.validator.internal.metadata.descriptor.ConstraintDescriptorImpl;
 import org.hibernate.validator.internal.metadata.descriptor.ExecutableDescriptorImpl;
 import org.hibernate.validator.internal.metadata.facets.Cascadable;
 import org.hibernate.validator.internal.metadata.location.ConstraintLocation.ConstraintLocationKind;
-import org.hibernate.validator.internal.metadata.raw.BeanConfiguration;
-import org.hibernate.validator.internal.metadata.raw.ConfigurationSource;
-import org.hibernate.validator.internal.metadata.raw.ConstrainedElement;
-import org.hibernate.validator.internal.metadata.raw.ConstrainedElement.ConstrainedElementKind;
-import org.hibernate.validator.internal.metadata.raw.ConstrainedExecutable;
-import org.hibernate.validator.internal.metadata.raw.ConstrainedField;
-import org.hibernate.validator.internal.metadata.raw.ConstrainedType;
-import org.hibernate.validator.internal.properties.Callable;
 import org.hibernate.validator.internal.util.CollectionHelper;
 import org.hibernate.validator.internal.util.ExecutableHelper;
-import org.hibernate.validator.internal.util.ExecutableParameterNameProvider;
-import org.hibernate.validator.internal.util.StringHelper;
-import org.hibernate.validator.internal.util.TypeResolutionHelper;
 import org.hibernate.validator.internal.util.classhierarchy.ClassHierarchyHelper;
 import org.hibernate.validator.internal.util.classhierarchy.Filters;
 import org.hibernate.validator.internal.util.logging.Log;
@@ -68,6 +54,7 @@ import org.hibernate.validator.spi.group.DefaultGroupSequenceProvider;
  * @author Kevin Pollet &lt;kevin.pollet@serli.com&gt; (C) 2011 SERLI
  * @author Chris Beckey &lt;cbeckey@paypal.com&gt;
  * @author Guillaume Smet
+ * @author Marko Bekhta
  */
 public final class BeanMetaDataImpl<T> implements BeanMetaData<T> {
 
@@ -267,7 +254,7 @@ public final class BeanMetaDataImpl<T> implements BeanMetaData<T> {
 		BeanDescriptor beanDescriptor = this.beanDescriptor;
 
 		if ( beanDescriptor == null ) {
-			synchronized ( this ) {
+			synchronized (this) {
 				beanDescriptor = this.beanDescriptor;
 
 				if ( beanDescriptor == null ) {
@@ -349,9 +336,9 @@ public final class BeanMetaDataImpl<T> implements BeanMetaData<T> {
 		if ( hasDefaultGroupSequenceProvider() ) {
 			List<Class<?>> providerDefaultGroupSequence = defaultGroupSequenceProvider.getValidationGroups( beanState );
 			return validationOrderGenerator.getDefaultValidationOrder(
-						beanClass,
-						getValidDefaultGroupSequence( beanClass, providerDefaultGroupSequence )
-					)
+					beanClass,
+					getValidDefaultGroupSequence( beanClass, providerDefaultGroupSequence )
+			)
 					.getSequenceIterator();
 		}
 		else {
@@ -435,9 +422,9 @@ public final class BeanMetaDataImpl<T> implements BeanMetaData<T> {
 			if ( executableMetaData.getKind() == ElementKind.METHOD
 					&& executableMetaData.isConstrained() ) {
 				ExecutableDescriptorImpl descriptor = executableMetaData.asDescriptor(
-								defaultGroupSequenceIsRedefined,
-								resolvedDefaultGroupSequence
-						);
+						defaultGroupSequenceIsRedefined,
+						resolvedDefaultGroupSequence
+				);
 
 				for ( String signature : executableMetaData.getSignatures() ) {
 					constrainedMethodDescriptors.put( signature, descriptor );
@@ -569,303 +556,6 @@ public final class BeanMetaDataImpl<T> implements BeanMetaData<T> {
 				+ ", constraintCount=" + getMetaConstraints().size()
 				+ ", cascadedPropertiesCount=" + cascadedProperties.size()
 				+ ", defaultGroupSequence=" + getDefaultGroupSequence( null ) + '}';
-	}
-
-	public static class PropertyHolderBeanMetaDataBuilder<T> {
-
-	}
-
-	public static class BeanMetaDataBuilder<T> {
-
-		private final ConstraintHelper constraintHelper;
-		private final ValidationOrderGenerator validationOrderGenerator;
-		private final Class<T> beanClass;
-		private final Set<BuilderDelegate> builders = newHashSet();
-		private final ExecutableHelper executableHelper;
-		private final TypeResolutionHelper typeResolutionHelper;
-		private final ValueExtractorManager valueExtractorManager;
-		private final ExecutableParameterNameProvider parameterNameProvider;
-		private final MethodValidationConfiguration methodValidationConfiguration;
-
-		private ConfigurationSource sequenceSource;
-		private ConfigurationSource providerSource;
-		private List<Class<?>> defaultGroupSequence;
-		private DefaultGroupSequenceProvider<? super T> defaultGroupSequenceProvider;
-
-
-		private BeanMetaDataBuilder(
-				ConstraintHelper constraintHelper,
-				ExecutableHelper executableHelper,
-				TypeResolutionHelper typeResolutionHelper,
-				ValueExtractorManager valueExtractorManager,
-				ExecutableParameterNameProvider parameterNameProvider,
-				ValidationOrderGenerator validationOrderGenerator,
-				Class<T> beanClass,
-				MethodValidationConfiguration methodValidationConfiguration) {
-			this.beanClass = beanClass;
-			this.constraintHelper = constraintHelper;
-			this.validationOrderGenerator = validationOrderGenerator;
-			this.executableHelper = executableHelper;
-			this.typeResolutionHelper = typeResolutionHelper;
-			this.valueExtractorManager = valueExtractorManager;
-			this.parameterNameProvider = parameterNameProvider;
-			this.methodValidationConfiguration = methodValidationConfiguration;
-		}
-
-		public static <T> BeanMetaDataBuilder<T> getInstance(
-				ConstraintHelper constraintHelper,
-				ExecutableHelper executableHelper,
-				TypeResolutionHelper typeResolutionHelper,
-				ValueExtractorManager valueExtractorManager,
-				ExecutableParameterNameProvider parameterNameProvider,
-				ValidationOrderGenerator validationOrderGenerator,
-				Class<T> beanClass,
-				MethodValidationConfiguration methodValidationConfiguration) {
-			return new BeanMetaDataBuilder<>(
-					constraintHelper,
-					executableHelper,
-					typeResolutionHelper,
-					valueExtractorManager,
-					parameterNameProvider,
-					validationOrderGenerator,
-					beanClass,
-					methodValidationConfiguration );
-		}
-
-		public void add(BeanConfiguration<? super T> configuration) {
-			if ( configuration.getBeanClass().equals( beanClass ) ) {
-				if ( configuration.getDefaultGroupSequence() != null
-						&& ( sequenceSource == null || configuration.getSource()
-						.getPriority() >= sequenceSource.getPriority() ) ) {
-
-					sequenceSource = configuration.getSource();
-					defaultGroupSequence = configuration.getDefaultGroupSequence();
-				}
-
-				if ( configuration.getDefaultGroupSequenceProvider() != null
-						&& ( providerSource == null || configuration.getSource()
-						.getPriority() >= providerSource.getPriority() ) ) {
-
-					providerSource = configuration.getSource();
-					defaultGroupSequenceProvider = configuration.getDefaultGroupSequenceProvider();
-				}
-			}
-
-			for ( ConstrainedElement constrainedElement : configuration.getConstrainedElements() ) {
-				addMetaDataToBuilder( constrainedElement, builders );
-			}
-		}
-
-		private void addMetaDataToBuilder(ConstrainedElement constrainableElement, Set<BuilderDelegate> builders) {
-			for ( BuilderDelegate builder : builders ) {
-				boolean foundBuilder = builder.add( constrainableElement );
-
-				if ( foundBuilder ) {
-					return;
-				}
-			}
-
-			builders.add(
-					new BuilderDelegate(
-							beanClass,
-							constrainableElement,
-							constraintHelper,
-							executableHelper,
-							typeResolutionHelper,
-							valueExtractorManager,
-							parameterNameProvider,
-							methodValidationConfiguration
-					)
-			);
-		}
-
-		public BeanMetaDataImpl<T> build() {
-			Set<ConstraintMetaData> aggregatedElements = newHashSet();
-
-			for ( BuilderDelegate builder : builders ) {
-				aggregatedElements.addAll( builder.build() );
-			}
-
-			return new BeanMetaDataImpl<>(
-					beanClass,
-					defaultGroupSequence,
-					defaultGroupSequenceProvider,
-					aggregatedElements,
-					validationOrderGenerator
-			);
-		}
-	}
-
-	private static class BuilderDelegate {
-		private final Class<?> beanClass;
-		private final ConstrainedElement constrainedElement;
-		private final ConstraintHelper constraintHelper;
-		private final ExecutableHelper executableHelper;
-		private final TypeResolutionHelper typeResolutionHelper;
-		private final ValueExtractorManager valueExtractorManager;
-		private final ExecutableParameterNameProvider parameterNameProvider;
-		private MetaDataBuilder metaDataBuilder;
-		private ExecutableMetaData.Builder methodBuilder;
-		private final MethodValidationConfiguration methodValidationConfiguration;
-		private final int hashCode;
-
-		public BuilderDelegate(
-				Class<?> beanClass,
-				ConstrainedElement constrainedElement,
-				ConstraintHelper constraintHelper,
-				ExecutableHelper executableHelper,
-				TypeResolutionHelper typeResolutionHelper,
-				ValueExtractorManager valueExtractorManager,
-				ExecutableParameterNameProvider parameterNameProvider,
-				MethodValidationConfiguration methodValidationConfiguration
-		) {
-			this.beanClass = beanClass;
-			this.constrainedElement = constrainedElement;
-			this.constraintHelper = constraintHelper;
-			this.executableHelper = executableHelper;
-			this.typeResolutionHelper = typeResolutionHelper;
-			this.valueExtractorManager = valueExtractorManager;
-			this.parameterNameProvider = parameterNameProvider;
-			this.methodValidationConfiguration = methodValidationConfiguration;
-
-			switch ( constrainedElement.getKind() ) {
-				case FIELD:
-					ConstrainedField constrainedField = (ConstrainedField) constrainedElement;
-					metaDataBuilder = new PropertyMetaData.Builder(
-							beanClass,
-							constrainedField,
-							constraintHelper,
-							typeResolutionHelper,
-							valueExtractorManager
-					);
-					break;
-				case CONSTRUCTOR:
-				case METHOD:
-				case GETTER:
-					ConstrainedExecutable constrainedExecutable = (ConstrainedExecutable) constrainedElement;
-					Callable callable = constrainedExecutable.getCallable();
-
-					// HV-890 Not adding meta-data for private super-type methods to the method meta-data of this bean;
-					// It is not needed and it may conflict with sub-type methods of the same signature
-					if ( !callable.isPrivate() || beanClass == callable.getDeclaringClass() ) {
-						methodBuilder = new ExecutableMetaData.Builder(
-								beanClass,
-								constrainedExecutable,
-								constraintHelper,
-								executableHelper,
-								typeResolutionHelper,
-								valueExtractorManager,
-								parameterNameProvider,
-								methodValidationConfiguration
-						);
-					}
-
-					if ( constrainedElement.getKind() == ConstrainedElementKind.GETTER ) {
-						metaDataBuilder = new PropertyMetaData.Builder(
-								beanClass,
-								constrainedExecutable,
-								constraintHelper,
-								typeResolutionHelper,
-								valueExtractorManager
-						);
-					}
-					break;
-				case TYPE:
-					ConstrainedType constrainedType = (ConstrainedType) constrainedElement;
-					metaDataBuilder = new ClassMetaData.Builder(
-							beanClass,
-							constrainedType,
-							constraintHelper,
-							typeResolutionHelper,
-							valueExtractorManager
-					);
-					break;
-				default:
-					throw new IllegalStateException(
-							StringHelper.format( "Constrained element kind '%1$s' not supported here.", constrainedElement.getKind() ) );
-			}
-
-			this.hashCode = buildHashCode();
-		}
-
-		public boolean add(ConstrainedElement constrainedElement) {
-			boolean added = false;
-
-			if ( methodBuilder != null && methodBuilder.accepts( constrainedElement ) ) {
-				methodBuilder.add( constrainedElement );
-				added = true;
-			}
-
-			if ( metaDataBuilder != null && metaDataBuilder.accepts( constrainedElement ) ) {
-				metaDataBuilder.add( constrainedElement );
-
-				if ( !added && constrainedElement.getKind().isMethod() && methodBuilder == null ) {
-					ConstrainedExecutable constrainedMethod = (ConstrainedExecutable) constrainedElement;
-					methodBuilder = new ExecutableMetaData.Builder(
-							beanClass,
-							constrainedMethod,
-							constraintHelper,
-							executableHelper,
-							typeResolutionHelper,
-							valueExtractorManager,
-							parameterNameProvider,
-							methodValidationConfiguration
-					);
-				}
-
-				added = true;
-			}
-
-			return added;
-		}
-
-		public Set<ConstraintMetaData> build() {
-			Set<ConstraintMetaData> metaDataSet = newHashSet();
-
-			if ( metaDataBuilder != null ) {
-				metaDataSet.add( metaDataBuilder.build() );
-			}
-
-			if ( methodBuilder != null ) {
-				metaDataSet.add( methodBuilder.build() );
-			}
-
-			return metaDataSet;
-		}
-
-		@Override
-		public int hashCode() {
-			return hashCode;
-		}
-
-		private int buildHashCode() {
-			final int prime = 31;
-			int result = 1;
-			result = prime * result + beanClass.hashCode();
-			result = prime * result + constrainedElement.hashCode();
-			return result;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if ( this == obj ) {
-				return true;
-			}
-			if ( !super.equals( obj ) ) {
-				return false;
-			}
-			if ( getClass() != obj.getClass() ) {
-				return false;
-			}
-			BuilderDelegate other = (BuilderDelegate) obj;
-			if ( !beanClass.equals( other.beanClass ) ) {
-				return false;
-			}
-			if ( !constrainedElement.equals( other.constrainedElement ) ) {
-				return false;
-			}
-			return true;
-		}
 	}
 
 	/**
